% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxon.R
\name{addTaxon}
\alias{addTaxon}
\title{Add prior information using 'features' as key}
\usage{
addTaxon(x, prior_info, reformat = TRUE)
}
\arguments{
\item{x}{A data frame with the key column labeled as `features`}

\item{prior_info}{}

\item{reformat}{Logical. Under the default (`TRUE`), any arbitrary 
delimiters will be removed and the rest will be concatenated.}
}
\description{
Add prior information using 'features' as key
}
