% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTaxonomicPrefix.R
\name{addTaxonomicPrefixes}
\alias{addTaxonomicPrefixes}
\title{Function to add taxonomic level prefixes}
\usage{
addTaxonomicPrefixes(taxonomic_names, current_delim)
}
\arguments{
\item{taxonomic_names}{A character vector}

\item{current_delim}{A character(1). Regular expression of the current delimiter}
}
\description{
Function to add taxonomic level prefixes
}
\examples{
taxonomic_name <- "Bacteria.Firmicutes.Bacilli.Lactobacillales.Streptococcaceae"
addTaxonomicPrefixes(taxonomic_name, "\\\\.")

}
