% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatInput.R
\name{formatInput}
\alias{formatInput}
\title{Format metagenomic data for different biomarker discovery tools}
\usage{
formatInput(dat, format_to, addOTU = TRUE)
}
\arguments{
\item{dat}{A microbiome abundance data. It can be a data frame (an input
format for LEfSe's `run_lefse.py` function), `SummarizedExperiment`, or
`phyloseq` object.}

\item{format_to}{The target format. Available options are
`c("SummarizedExperiment", "phyloseq")`.}

\item{addOTU}{Logical. If `TRUE`, feature name/number is kept with the
`OTU_` prefix. If `FALSE`, only the taxonomic levels are used.}
}
\description{
Input format conversion across `SummarizedExperiment` and `phyloseq` objects.
}
\examples{
formatInput(kostic_crc_small, format_to = "SummarizedExperiment")

}
