% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxNameToTable.R
\name{taxNameToTable}
\alias{taxNameToTable}
\title{Split the taxonomic label in character into the tax_table format}
\usage{
taxNameToTable(x, prefix = TRUE)
}
\arguments{
\item{x}{A character vector of taxonomic names}

\item{prefix}{Logical. If it's `TRUE`, there is a prefix showing the
taxonomic levels.}
}
\description{
Split the taxonomic label in character into the tax_table format
}
