% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxTableToName.R
\name{taxTableToName}
\alias{taxTableToName}
\title{Combine tax_table information into a character vector}
\usage{
taxTableToName(taxTable, delim = "|")
}
\arguments{
\item{taxTable}{A `taxonomyTable` object from phyloseq.}

\item{delim}{A character(1). Delimeter to separate taxon levels.}
}
\description{
Combine tax_table information into a character vector
}
\examples{
data(kostic_crc_small)
taxTableToName(tax_table(kostic_crc_small))

}
