% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeVennDiagram.R
\name{threeVennDiagram}
\alias{threeVennDiagram}
\title{Create input for Venn Diagram plot by eulerr package}
\usage{
threeVennDiagram(lefse_taxa, lefser_taxa, mm_taxa)
}
\arguments{
\item{lefse_taxa}{A character vector.}

\item{lefser_taxa}{A character vector.}

\item{mm_taxa}{A character vector.}
}
\value{
An `euler` object ready for `plot` function.
}
\description{
Create input for Venn Diagram plot by eulerr package
}
