% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeVennDiagram.R
\name{twoVennDiagram}
\alias{twoVennDiagram}
\title{Create input for Venn Diagram plot by eulerr package}
\usage{
twoVennDiagram(input1, input2, name1 = "A", name2 = "B")
}
\arguments{
\item{input1}{A character vector.}

\item{input2}{A character vector.}

\item{name1}{A character(1).}

\item{name2}{A character(1).}
}
\value{
An `euler` object ready for `plot` function.
}
\description{
Create input for Venn Diagram plot by eulerr package
}
